+++ Runtime/Utils/UrlUtils.cs
@@ -1,10 +1,8 @@
-﻿// Copyright (c) 2023 AccelByte Inc. All Rights Reserved.
+﻿// Copyright (c) 2023 - 2026 AccelByte Inc. All Rights Reserved.
 // This is licensed software from AccelByte Inc, for limitations
 // and restrictions contact your company contract manager.
 
 using System.Text.RegularExpressions;
-using AccelByte.Core;
-
 
 namespace AccelByte.Utils
 {
@@ -18,7 +16,11 @@ namespace AccelByte.Utils
                 var regexStr = "^https?|wss?:\\/\\/(?:www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b(?:[-a-zA-Z0-9()@:%_\\+.~#?&\\/=]*)$";
                 if (Regex.IsMatch(baseURL, regexStr))
                 {
-                    retval = baseURL.TrimEnd('/');
+                    retval = baseURL;
+                    while (retval.EndsWith('/'))
+                    {
+                        retval = retval.TrimEnd('/');
+                    }
                 }
                 else
                 {